# specs/004-institution-iam-auth-integration/contracts/api-openapi.yaml
openapi: 3.0.3
info:
  title: Clinica Digital IAM API
  version: 0.3.0
  description: Contratos de IAM nativo in-app, registro institucional, sessao opaca, contexto do Shell e validacao RNDS para Organization, Location, Practitioner e PractitionerRole.
servers:
  - url: http://localhost:8080
paths:
...
  /api/admin/locations:
    get:
      summary: Lista locations administrativas do tenant autenticado
      operationId: listAdminLocations
      responses:
        '200':
          description: Lista de locations retornada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminLocationResponse'
        '401':
          description: Sessao invalida ou expirada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '403':
          description: Somente admin (profile 10)
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
    post:
      summary: Cria location administrativa no tenant autenticado
      operationId: createAdminLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminLocationRequest'
      responses:
        '201':
          description: Location criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLocationResponse'
        '400':
          description: Payload invalido
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Organizacao nao encontrada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /api/admin/locations/{locationId}:
    put:
      summary: Atualiza location administrativa no tenant autenticado
      operationId: updateAdminLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminLocationRequest'
      responses:
        '200':
          description: Location atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLocationResponse'
        '404':
          description: Location nao encontrada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /api/admin/locations/{locationId}/deactivate:
    post:
      summary: Desativa location administrativa no tenant autenticado
      operationId: deactivateAdminLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Location desativada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLocationResponse'
        '404':
          description: Location nao encontrada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /api/admin/practitioner-roles:
    get:
      summary: Lista lotacoes administrativas do tenant autenticado
      operationId: listAdminPractitionerRoles
      responses:
        '200':
          description: Lista de lotacoes retornada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminPractitionerRoleResponse'
        '401':
          description: Sessao invalida ou expirada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '403':
          description: Somente admin (profile 10)
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
    post:
      summary: Cria lotacao administrativa no tenant autenticado
      operationId: createAdminPractitionerRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdminPractitionerRoleRequest'
      responses:
        '201':
          description: Lotacao criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPractitionerRoleResponse'
        '400':
          description: Payload invalido
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
        '404':
          description: Dependencia nao encontrada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /api/admin/practitioner-roles/{roleId}:
    put:
      summary: Atualiza lotacao administrativa no tenant autenticado
      operationId: updateAdminPractitionerRole
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdminPractitionerRoleRequest'
      responses:
        '200':
          description: Lotacao atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPractitionerRoleResponse'
        '404':
          description: Lotacao nao encontrada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
  /api/admin/practitioner-roles/{roleId}/deactivate:
    post:
      summary: Desativa lotacao administrativa no tenant autenticado
      operationId: deactivateAdminPractitionerRole
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lotacao desativada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPractitionerRoleResponse'
        '404':
          description: Lotacao nao encontrada
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/OperationOutcome'
...
components:
  schemas:
...
    CreateAdminLocationRequest:
      type: object
      required: [organizationId, displayName]
      properties:
        organizationId:
          type: string
          format: uuid
        displayName:
          type: string
        fhirName:
          type: string
        fhirStatus:
          type: string
        fhirMode:
          type: string
    UpdateAdminLocationRequest:
      type: object
      properties:
        displayName:
          type: string
        fhirName:
          type: string
        fhirStatus:
          type: string
        fhirMode:
          type: string
        accountActive:
          type: boolean
    AdminLocationResponse:
      type: object
      required: [id, tenantId, organizationId, displayName, fhirName, fhirStatus, fhirMode, accountActive]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        displayName:
          type: string
        fhirName:
          type: string
        fhirStatus:
          type: string
        fhirMode:
          type: string
        accountActive:
          type: boolean
    CreateAdminPractitionerRoleRequest:
      type: object
      required: [organizationId, locationId, practitionerId, roleCode]
      properties:
        organizationId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        roleCode:
          type: string
        primaryRole:
          type: boolean
        fhirCodeJson:
          type: string
        fhirSpecialtyJson:
          type: string
        fhirTelecomJson:
          type: string
        fhirAvailableTimeJson:
          type: string
    UpdateAdminPractitionerRoleRequest:
      type: object
      properties:
        roleCode:
          type: string
        active:
          type: boolean
        primaryRole:
          type: boolean
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        fhirCodeJson:
          type: string
        fhirSpecialtyJson:
          type: string
        fhirTelecomJson:
          type: string
        fhirAvailableTimeJson:
          type: string
    AdminPractitionerRoleResponse:
      type: object
      required: [id, tenantId, organizationId, locationId, practitionerId, roleCode, active, primaryRole]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        locationId:
          type: string
          format: uuid
        practitionerId:
          type: string
          format: uuid
        roleCode:
          type: string
        active:
          type: boolean
        primaryRole:
          type: boolean
        periodStart:
          type: string
          format: date-time
          nullable: true
        periodEnd:
          type: string
          format: date-time
          nullable: true
        fhirCodeJson:
          type: string
          nullable: true
        fhirSpecialtyJson:
          type: string
          nullable: true
        fhirTelecomJson:
          type: string
          nullable: true
        fhirAvailableTimeJson:
          type: string
          nullable: true
...
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          minItems: 1
          items:
            type: object
            required: [severity, code, details]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
              details:
                type: object
                required: [text]
                properties:
                  text:
                    type: string
              diagnostics:
                type: string